appveyor.yml

+++ src/segyio/segy.c
@@ -576,7 +576,7 @@ int segy_sample_indexes(FILE* fp, double* buf, double t0, size_t count) {
         return err;
     }
 
-    for (int i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         buf[i] = t0 + i * dt;
     }
 
@@ -1088,7 +1088,7 @@ int segy_write_textheader( FILE* fp, unsigned int pos, const char* buf ) {
     return SEGY_OK;
 }
 
-unsigned int segy_textheader_size() {
+int segy_textheader_size() {
     return SEGY_TEXT_HEADER_SIZE + 1;
 }
